Move code from service/common/telemetry to the existing telemetry pkg
