+++ internal/pkg/telemetry/span.go
